% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_general.R
\name{calculate_landscape_matrix}
\alias{calculate_landscape_matrix}
\title{Calculate the landscape/connectivity matrix from given metapopulation parameters}
\usage{
calculate_landscape_matrix(
  distance,
  patch_area,
  dispersal_kernel,
  dispersal_rate = 1,
  extinction_rate = 0.5,
  self_regulation = TRUE
)
}
\arguments{
\item{distance}{a matrix containing pairwise distances between patches}

\item{patch_area}{a vector containing the patch areas/qualities}

\item{dispersal_kernel}{which dispersal kernel to use}

\item{dispersal_rate}{area dependence of dispersal}

\item{extinction_rate}{area dependence of extinction}

\item{self_regulation}{whether self-colonization exists}
}
\value{
landscape/connectivity matrix
}
\description{
Calculate the landscape/connectivity matrix from given metapopulation parameters
}
