% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_general.R
\name{calculate_metapopulation_capacity}
\alias{calculate_metapopulation_capacity}
\title{Calculate metapopulation capacity and patch importance}
\usage{
calculate_metapopulation_capacity(m, patch_mc = FALSE)
}
\arguments{
\item{m}{landscape/connectivity matrix}

\item{patch_mc}{whether to compute patch importance}
}
\value{
indicators of metapopulation persistence
}
\description{
Calculate metapopulation capacity and patch importance
}
