% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_prediction.R
\name{estimate_asymptotic_capacity}
\alias{estimate_asymptotic_capacity}
\title{Analytic-based inference of metapopulation capacity}
\usage{
estimate_asymptotic_capacity(m, total_num)
}
\arguments{
\item{m}{landscape/connectivity matrix}

\item{total_num}{total number of patches in a metapopulation}
}
\value{
predicted metapopulation capacity
}
\description{
Analytic-based inference of metapopulation capacity
}
