% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_prediction.R
\name{estimate_offdigonal_mean}
\alias{estimate_offdigonal_mean}
\title{Calculate the mean of off-diagonal elements in a matrix}
\usage{
estimate_offdigonal_mean(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
mean of off-diagonal elements
}
\description{
Calculate the mean of off-diagonal elements in a matrix
}
