% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_prediction.R
\name{estimate_offdigonal_variance}
\alias{estimate_offdigonal_variance}
\title{Calculate the variance of off-diagonal elements in a matrix}
\usage{
estimate_offdigonal_variance(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
variance of off-diagonal elements
}
\description{
Calculate the variance of off-diagonal elements in a matrix
}
