% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_prediction.R
\name{estimate_ratio_area}
\alias{estimate_ratio_area}
\title{Estimate the ratio of dependency}
\usage{
estimate_ratio_area(A0, self_regulation)
}
\arguments{
\item{A0}{area matrix (A_i^e A_j^omega)}

\item{self_regulation}{whether self-colonization exists}
}
\value{
ratio of dependency
}
\description{
Estimate the ratio of dependency
}
