% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_nullmodel.R
\name{generate_dispersal_kernel}
\alias{generate_dispersal_kernel}
\title{Generate dispersal kernels}
\usage{
generate_dispersal_kernel(kernel_type, para, ...)
}
\arguments{
\item{kernel_type}{which kernel to use}

\item{para}{parameter in the kernel}

\item{...}{other arguments passed to methods}
}
\value{
a function of dispersal kernel
}
\description{
Generate dispersal kernels
}
