% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_nullmodel.R
\name{generate_metapopulation}
\alias{generate_metapopulation}
\title{Generate metapopulation}
\usage{
generate_metapopulation(
  patch_num,
  kernel_type,
  charactersitic_distance,
  location_distribution,
  self_regulation,
  dispersal_rate,
  extinction_rate,
  area_type,
  area_para,
  ...
)
}
\arguments{
\item{patch_num}{Number of patches}

\item{kernel_type}{which dispersal kernel to use}

\item{charactersitic_distance}{characteristic dispersal distance}

\item{location_distribution}{how patches are spatially distributed}

\item{self_regulation}{whether self-colonization exists}

\item{dispersal_rate}{area dependence of dispersal}

\item{extinction_rate}{area dependence of extinction}

\item{area_type}{distribution of patch area/quality}

\item{area_para}{parameter for area/quality distribution}

\item{...}{other arguments passed to methods}
}
\value{
a list of metapopulation parameters
}
\description{
Generate metapopulation
}
