% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_nullmodel.R
\name{generate_patch_area}
\alias{generate_patch_area}
\title{Generate patch areas/qualities}
\usage{
generate_patch_area(patch_num, area_type, area_para)
}
\arguments{
\item{patch_num}{number of patches}

\item{area_type}{distribution of patch area/quality}

\item{area_para}{parameter for area/quality distribution}
}
\value{
a vector of patch areas/qualities
}
\description{
Generate patch areas/qualities
}
