% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_nullmodel.R
\name{generate_patch_location}
\alias{generate_patch_location}
\title{Generate spatial distributions of patches}
\usage{
generate_patch_location(
  patch_type,
  patch_num = 100,
  maxRange = 1,
  ncluster = 5
)
}
\arguments{
\item{patch_type}{which spatial distribution to use}

\item{patch_num}{number of patches}

\item{maxRange}{parameter for spatial interactions}

\item{ncluster}{number of spatial clusters}
}
\value{
a vector of patch areas/qualities
}
\description{
Generate spatial distributions of patches
}
