% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_prediction.R
\name{predict_capacity}
\alias{predict_capacity}
\title{Regression-based inference of metapopulation capacity}
\usage{
predict_capacity(df, method)
}
\arguments{
\item{df}{a tibble of sub-metapopulations and their capacities}

\item{method}{regression method}
}
\value{
predicted metapopulation capacity
}
\description{
Regression-based inference of metapopulation capacity
}
