% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox_subsampling.R
\name{sample_sub_metapopulation_mc}
\alias{sample_sub_metapopulation_mc}
\title{Calculate the patch importance of sub-metapopulations}
\usage{
sample_sub_metapopulation_mc(
  M,
  subpatch_num_min = 2,
  subpatch_num_max = patch_num/2,
  Nrand_num = 20
)
}
\arguments{
\item{M}{landscape/connectivity matrix}

\item{subpatch_num_min}{minimum number of samples}

\item{subpatch_num_max}{maximum number of samples}

\item{Nrand_num}{number of replicates with a given number of sampled patches}
}
\value{
patch importance for sub-metapopulations
}
\description{
Calculate the patch importance of sub-metapopulations
}
